+++ android/app/src/main/kotlin/com/example/camera_x_demo/MainActivity.kt
@@ -1,5 +1,0 @@
-package com.example.camera_x_demo
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity : FlutterActivity()